requirements.txt
fastapi
fastapi[cors]
uvicorn[standard]
python-multipart
torch
torchvision
timm
requests
Pillow
pydantic-settings
python-dotenv
google-generativeai
scipy
